crypto: rc4: optimize calculations
Calculate SBox in RC4 constructor rather than do it every time on message
encryption or decription. In this case both IV and KEY as well as KBox
are not needed since they are just inputs for SBox. So do not include
them into RC4 structure. However size params are needed as long as we
use stream_cipher traits. So add two PhantomData internal handlers.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>